server/requirements.txt
algoliasearch-django>=2.0,<3.0
django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.5.1
pyyaml
requests
django-cors-headers==4.6.0
gunicorn